+++ src/main/java/org/folio/service/processing/ranking/QueueItemHolisticRanker.java
@@ -23,7 +23,7 @@ public class QueueItemHolisticRanker implements QueueItemRanker {
   @Override
   public double score(
     DataImportQueueItem queueItem,
-    Map<String, Integer> tenantUsage
+    Map<String, Long> tenantUsage
   ) {
     return (
       ageRanker.score(queueItem, tenantUsage) +
