+++ src/main/java/org/folio/service/processing/ranking/QueueItemRanker.java
@@ -7,5 +7,5 @@ public interface QueueItemRanker {
   /**
    * Calculate the score for a queue item based on the class's configuration
    */
-  double score(DataImportQueueItem queueItem, Map<String, Integer> tenantUsage);
+  double score(DataImportQueueItem queueItem, Map<String, Long> tenantUsage);
 }
