+++ src/main/java/org/folio/service/processing/ranking/QueueItemSizeRanker.java
@@ -25,7 +25,7 @@ public class QueueItemSizeRanker implements QueueItemRanker {
   @Override
   public double score(
     DataImportQueueItem queueItem,
-    Map<String, Integer> tenantUsage
+    Map<String, Long> tenantUsage
   ) {
     return ScoreUtils.calculateUnboundedLogarithmicScore(
       queueItem.getOriginalSize(),
